+++ CHANGES.rst
@@ -6,8 +6,7 @@ Changelog
 4.1.28.2 (unreleased)
 ---------------------
 
-- Nothing changed yet.
-
+- Re-run `_removeBrokenAnnexes` from `Migrate_To_4105` in `Migrate_To_4110` in case some broken annexes are still found
 
 4.1.28.1 (2020-08-21)
 ---------------------